1.6

+++ resources/views/order/index.blade.php
@@ -11,9 +11,11 @@
         <tr>
             <th>Rb</th>
             <th>Korisnik</th>
-            <th>Taxi sluzba</th>
-            <th>Pocetna adresa</th>
-            <th>Zavrsna adresa</th>
+            <th>Taxi služba</th>
+            <th>Početna adresa</th>
+            <th>Završna adresa</th>
+            <th>Udaljenost</th>
+            <th>Cijena</th>
             <th>Datum i vrijeme</th>
             <th>Akcija</th>
         </tr>
@@ -21,14 +23,16 @@
         <tbody>
         @foreach($orders as $order)
             <tr>
-                <td>{{ $index }}</td>
+                <td><a href="{{ url('order/'.$order->id) }}">{{ $index }}</td>
                 <td>{{ $order->user }}</td>
                 <td>{{ $order->taxi }}</td>
                 <td>{{ $order->startAddress }}</td>
                 <td>{{ $order->endAddress }}</td>
+                <td>{{ $order->distance }} km</td>
+                <td>{{ $order->price }} kn</td>
                 <td>{{ $order->created_at }}</td>
-                {{ Form::open(['method'=>'DELETE','route'=>['order.destroy',$order->id]]) }}
-                <td>{{ Form::submit('Delete',['class'=>'btn btn-danger']) }}</td>
+                {{ Form::open(['method'=>'DELETE','route'=>['order.destroy',$order->id],'onsubmit'=>'return ConfirmDelete() ']) }}
+                <td>{{ Form::submit('Obriši',['class'=>'btn btn-danger']) }}</td>
                 {{ Form::close() }}
             </tr>
             {{--*/ $index++ /*--}}
